requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
websocket-client==1.6.4
gTTS==2.4.0 